backend/resumeparser/requirements.txt
flask==2.3.2
flask-cors==4.0.0
PyPDF2==3.0.1
sentence-transformers==2.2.2
scikit-learn==1.3.0
numpy==1.24.3
pdfminer.six==20221105
